fix user balance amount

+++ backend/src/calculate_share.py
@@ -1,5 +1,5 @@
 def calculate_each_share(expense_amount, expense_shared_by):
-    each_share_amount = int(expense_amount) / len(expense_shared_by.split(','))
+    each_share_amount = float(expense_amount) / len(expense_shared_by.split(','))
     return each_share_amount
